Refactor: move base classes into a separate file
The first of a series of refactoring moves that splits binding.py
into manageable chunks to make it easier to maintain "going forward"
