+++ opsramp/binding.py
@@ -2,6 +2,9 @@
 #
 # A minimal Python language binding for the OpsRamp REST API.
 #
+# binding.py
+# Defines the primary entry points for callers of this library.
+#
 # (c) Copyright 2019 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,8 +20,9 @@
 # limitations under the License.
 
 from __future__ import print_function
-import requests
 import base64
+
+from opsramp.base import ApiObject, ApiWrapper
 
 
 def connect(url, key, secret):
@@ -34,146 +38,6 @@ def connect(url, key, secret):
     auth_resp = ao.post(data=body)
     token = auth_resp['access_token']
     return Opsramp(url, token)
-
-
-class PathTracker(object):
-    def __init__(self):
-        self.reset()
-
-    def reset(self):
-        self.prefix = ''
-        self.stack = []
-
-    def __str__(self):
-        return '%s "%s" %s' % (str(type(self)), self.prefix, self.stack)
-
-    def clone(self):
-        new1 = PathTracker()
-        new1.prefix = self.prefix
-        new1.stack = self.stack
-        return new1
-
-    def cd(self, path):
-        # no support for '..' right now, maybe in the future
-        if path[0] == '/':
-            self.prefix = path
-        else:
-            self.prefix += '/' + path
-        self.prefix = self.prefix.strip('/')
-        return self.prefix
-
-    def pushd(self, path):
-        self.stack.append(self.prefix)
-        return self.cd(path)
-
-    def popd(self):
-        self.prefix = self.stack.pop()
-        return self.prefix
-
-    def fullpath(self, suffix=''):
-        if len(suffix) > 0 and suffix[0] == '/':
-            retval = suffix
-        else:
-            retval = ''
-            if len(self.prefix) > 0:
-                retval += '/' + self.prefix
-            if len(suffix) > 0:
-                retval += '/' + suffix
-        return retval
-
-
-class ApiObject(object):
-    def __init__(self, url, auth, tracker=None):
-        self.baseurl = url.rstrip('/')
-        self.auth = auth
-        if tracker:
-            self.tracker = tracker
-        else:
-            self.tracker = PathTracker()
-
-    def __str__(self):
-        return '%s "%s" "%s"' % (
-            str(type(self)), self.baseurl, self.tracker.fullpath()
-        )
-
-    def clone(self):
-        new1 = ApiObject(self.baseurl, self.auth, self.tracker.clone())
-        return new1
-
-    def cd(self, path):
-        return self.tracker.cd(path)
-
-    def pushd(self, path):
-        return self.tracker.pushd(path)
-
-    def popd(self):
-        return self.tracker.popd()
-
-    def chroot(self, suffix=''):
-        suffix = self.tracker.fullpath(suffix)
-        if suffix:
-            self.baseurl += suffix
-            self.tracker.reset()
-        return ''
-
-    def compute_url(self, suffix=''):
-        retval = self.baseurl
-        suffix = self.tracker.fullpath(suffix)
-        if suffix:
-            retval += suffix
-        return retval.rstrip('/')
-
-    def prep_headers(self, headers):
-        if not headers:
-            return self.auth
-        hdr = {}
-        hdr.update(self.auth)
-        hdr.update(headers)
-        return hdr
-
-    @staticmethod
-    def process_result(url, resp):
-        if resp.status_code != requests.codes.OK:
-            msg = '%s %s %s' % (resp, url, resp.content)
-            raise RuntimeError(msg)
-        try:
-            return resp.json()
-        except: # noqa
-            return resp.text
-
-    def get(self, suffix='', headers={}):
-        url = self.compute_url(suffix)
-        hdr = self.prep_headers(headers)
-        resp = requests.get(url, headers=hdr)
-        return self.process_result(url, resp)
-
-    def post(self, suffix='', headers={}, data=None, json=None):
-        url = self.compute_url(suffix)
-        hdr = self.prep_headers(headers)
-        resp = requests.post(url, headers=hdr, data=data, json=json)
-        return self.process_result(url, resp)
-
-    def put(self, suffix='', headers={}, data=None, json=None):
-        url = self.compute_url(suffix)
-        hdr = self.prep_headers(headers)
-        resp = requests.put(url, headers=hdr, data=data, json=json)
-        return self.process_result(url, resp)
-
-    def delete(self, suffix='', headers={}):
-        url = self.compute_url(suffix)
-        hdr = self.prep_headers(headers)
-        resp = requests.delete(url, headers=hdr)
-        return self.process_result(url, resp)
-
-
-class ApiWrapper(object):
-    def __init__(self, apiobject, suffix=''):
-        self.api = apiobject.clone()
-        if suffix:
-            self.api.chroot(suffix)
-
-    def __str__(self):
-        return '%s %s' % (str(type(self)), self.api)
 
 
 class Opsramp(ApiWrapper):
